MetaGPT/requirements.txt
python-dotenv
requests
typer
pytest>=7.0
